+++ src/Linna/TypedArrayObject/ArrayOfBooleans.php
@@ -17,7 +17,7 @@ use InvalidArgumentException;
 
 /**
  * Provide a way to create an array of boolean typed elements with php.
- * 
+ *
  * @link https://www.php.net/manual/en/functions.first_class_callable_syntax.php
  */
 class ArrayOfBooleans extends AbstractArray
@@ -33,14 +33,14 @@ class ArrayOfBooleans extends AbstractArray
      * Class Contructor.
      *
      * @param array<bool>  $input          Array of values, every value must be a <code>bool</code>.
-     * @param int          $flags          Flags to control the behaviour of the <code>ArrayObject</code> object, 
+     * @param int          $flags          Flags to control the behaviour of the <code>ArrayObject</code> object,
      *                                     see <code>ArrayObject</code> on php site.
-     * @param class-string $iterator_class Specify the class that will be used for iteration of the <code>ArrayObject</code> 
+     * @param class-string $iterator_class Specify the class that will be used for iteration of the <code>ArrayObject</code>
      *                                     object, the class must implement <code>ArrayIterator</code>.
      *
      * @throws InvalidArgumentException If elements in the optional array parameter aren't of the configured type.
      */
-    public function __construct(array $input = [], int $flags = 0, string $iterator_class = ArrayIterator::class)
+    public function __construct(array $input = [], int $flags = AbstractArray::NO_FLAGS, string $iterator_class = ArrayIterator::class)
     {
         // first argument is the php8.1 method to pass function reference as closure.
         // check https://www.php.net/manual/en/functions.first_class_callable_syntax.php